test: moved all github url tests together
